Added dynamic product list
The product list on index.html is now dynamic. There's also a lot of thought for how we will need to expand this in the future, so there's capabilities to have different kinds of products and to sort the list. It should be really easy to change to just about any situation. The links on the tents are also broken, as they cannot function until the dynamic product pages is completed.